pyproject.toml
[virtualenvs]
create = true
in-project = true


[tool.poetry]
name = "airy"
version = "0.1.0"
description = ""
authors = ["CraazzzyyFoxx <38073783+CraazzzyyFoxx@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"

psutil = "^5.9.0"
asyncpg = "^0.25.0"
Levenshtein = "^0.18.1"
uvloop = {version = "==0.16.0", platform="linux"}
aiodns = "~=3.0"
cchardet = "~=2.1"
Brotli = "~=1.0"
ciso8601 = "~=2.2"
tortoise-orm = "~=0.19.0"
lru_dict = "~=1.1.7"
aioredis = "~=2.0.0"
aerich = "~=0.6.3"
orjson = "~=3.6.4"
pytz = "~=2021.1"
pydantic = "~=1.9.0"
python-dotenv = "~=0.20.0"
tekore = "~=4.3.0"
fuzzywuzzy = "~=0.18.0"
yarl = "~=1.7.2"
deepdiff = "~=5.7.0"
pygount = "~=1.3.0"
attrs = "~=21.4.0"
#python-json-logger = "~=2.0.2"

hikari-lightbulb = "~=2.2.1"

#dateparser = {git = "https://github.com/scrapinghub/dateparser.git", tag="v1.1.1"}
hikari-miru = { git = "https://github.com/HyperGH/hikari-miru.git", branch = "cursed/branch" }
"lavacord.py" = { git = "https://github.com/CraazzzyyFoxx/lavacord.py" }


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
